+++ packages/sdk/src/crypto/noteEncryption.ts
@@ -1,17 +1,9 @@
-import {
-  CanonAddress,
-  EncryptedCanonAddress,
-  StealthAddress,
-  StealthAddressTrait,
-} from "./address";
-import { Note, Asset } from "../primitives";
+import { CanonAddress } from "./address";
+import { Note, NoteTrait } from "../primitives";
 import { EncryptedNote } from "../primitives/types";
-import { BabyJubJub, poseidonBN } from "@nocturne-xyz/crypto-utils";
-import { randomFr } from "./utils";
-import { decompressPoint, compressPoint } from "./pointCompression";
+import { BabyJubJub, HybridCipher } from "@nocturne-xyz/crypto-utils";
 
-const F = BabyJubJub.BaseField;
-const Fr = BabyJubJub.ScalarField;
+const cipher = new HybridCipher(BabyJubJub, 64);
 
 /**
  * Encrypt a note to be decrypted by a given receiver
@@ -24,26 +16,9 @@ const Fr = BabyJubJub.ScalarField;
  * The viewing key corresponding to the receiver's canonical address is the decryption key
  */
 export function encryptNote(receiver: CanonAddress, note: Note): EncryptedNote {
-  const r = randomFr();
-  const R = BabyJubJub.scalarMul(BabyJubJub.BasePoint, r);
-
-  const encryptedNonce = F.add(
-    poseidonBN([compressPoint(R)]),
-    F.reduce(note.nonce)
-  );
-  const encryptedValue = F.add(
-    poseidonBN([F.reduce(compressPoint(R) + F.One)]),
-    F.reduce(note.value)
-  );
-
-  return {
-    owner: StealthAddressTrait.compress(
-      StealthAddressTrait.randomize(note.owner)
-    ),
-    encappedKey: compressPoint(BabyJubJub.scalarMul(receiver, r)),
-    encryptedNonce,
-    encryptedValue,
-  };
+  const noteBytes = NoteTrait.serializeCompact(note);
+  const ciphertext = cipher.encrypt(noteBytes, receiver);
+  return ciphertext;
 }
 
 /**
@@ -57,68 +32,7 @@ export function encryptNote(receiver: CanonAddress, note: Note): EncryptedNote {
  * `vk` need not be the viewing key corresponding to the owner's canonical address. The decryption process
  * will work as long as `encryptedNote` was encrypted with `vk`'s corresponding `CanonicalAddress`
  */
-export function decryptNote(
-  owner: StealthAddress,
-  vk: bigint,
-  encryptedNote: EncryptedNote,
-  asset: Asset
-): Note {
-  let vkInv = Fr.inv(vk);
-  if (vkInv < BabyJubJub.PrimeSubgroupOrder) {
-    vkInv += BabyJubJub.PrimeSubgroupOrder;
-  }
-
-  const eR = decompressPoint(encryptedNote.encappedKey);
-  if (!eR) {
-    throw new Error("Invalid encapped key");
-  }
-
-  const R = BabyJubJub.scalarMul(eR, vkInv);
-  const nonce = F.sub(
-    F.reduce(encryptedNote.encryptedNonce),
-    F.reduce(poseidonBN([compressPoint(R)]))
-  );
-
-  const value = F.sub(
-    F.reduce(encryptedNote.encryptedValue),
-    F.reduce(poseidonBN([F.reduce(compressPoint(R) + 1n)]))
-  );
-
-  return {
-    owner,
-    nonce,
-    asset,
-    value,
-  };
-}
-
-// ElGamal encryption using receiver's canonical address as the public key (vk the private key)
-export function encryptCanonAddr(
-  plaintext: CanonAddress,
-  pubkey: CanonAddress,
-  nonce: bigint
-): EncryptedCanonAddress {
-  const s = BabyJubJub.scalarMul(pubkey, nonce);
-  const c1 = BabyJubJub.scalarMul(BabyJubJub.BasePoint, nonce);
-  const c2 = BabyJubJub.add(plaintext, s);
-
-  return {
-    c1: compressPoint(c1),
-    c2: compressPoint(c2),
-  };
-}
-
-export function decryptCanonAddr(
-  ciphertext: EncryptedCanonAddress,
-  vk: bigint
-): CanonAddress {
-  const c1 = decompressPoint(ciphertext.c1);
-  const c2 = decompressPoint(ciphertext.c2);
-
-  if (!c1 || !c2) {
-    throw new Error("Invalid ciphertext");
-  }
-
-  const sInv = BabyJubJub.scalarMul(c1, Fr.neg(vk));
-  return BabyJubJub.add(c2, sInv);
+export function decryptNote(vk: bigint, encryptedNote: EncryptedNote): Note {
+  const noteBytes = cipher.decrypt(encryptedNote, vk);
+  return NoteTrait.deserializeCompact(noteBytes);
 }