+++ packages/sdk/src/primitives/note.ts
@@ -4,6 +4,7 @@ import { poseidonBN } from "@nocturne-xyz/crypto-utils";
 import { sha256 } from "js-sha256";
 import { bigintToBEPadded } from "../utils";
 import { EncryptedNote } from "./types";
+import { bigintFromBEBytes } from "../utils/bits";
 
 export interface Note {
   owner: StealthAddress;
@@ -32,6 +33,23 @@ export interface EncodedNote {
   encodedAssetId: bigint;
   value: bigint;
 }
+
+// compact note byte-serialization format (181 bytes)
+// all bigints are represented in big-endian encoding
+// 1. owner: 64 bytes (h1Compressed: 32 bytes, h2Compressed: 32 bytes)
+// 2. asset: 53 bytes (AssetTrait.serializeCompact(asset))
+// 3. nonce: 32 bytes
+// 4. value: 32 bytes
+const COMPACT_SERIALIZE_OFFSETS = {
+  owner: {
+    h1: 0,
+    h2: 32,
+  },
+  asset: 64,
+  nonce: 117,
+  value: 149,
+  end: 181,
+};
 
 export class NoteTrait {
   static toCommitment<N extends Note>(note: N): bigint {
@@ -124,6 +142,60 @@ export class NoteTrait {
   }
 
   static toNote<N extends Note>({ owner, nonce, asset, value }: N): Note {
+    return { owner, nonce, asset, value };
+  }
+
+  // see above for serialization format
+  static serializeCompact<N extends Note>(note: N): Uint8Array {
+    const buf = new Uint8Array(180);
+
+    const { owner, nonce, asset, value } = note;
+    const { h1, h2 } = StealthAddressTrait.compress(owner);
+
+    buf.set(bigintToBEPadded(h1, 32), COMPACT_SERIALIZE_OFFSETS.owner.h1);
+    buf.set(bigintToBEPadded(h2, 32), COMPACT_SERIALIZE_OFFSETS.owner.h2);
+    buf.set(
+      AssetTrait.serializeCompact(asset),
+      COMPACT_SERIALIZE_OFFSETS.asset
+    );
+    buf.set(bigintToBEPadded(nonce, 32), COMPACT_SERIALIZE_OFFSETS.nonce);
+    buf.set(bigintToBEPadded(value, 32), COMPACT_SERIALIZE_OFFSETS.value);
+
+    return buf;
+  }
+
+  static deserializeCompact(buf: Uint8Array): Note {
+    const h1 = bigintFromBEBytes(
+      buf.slice(
+        COMPACT_SERIALIZE_OFFSETS.owner.h1,
+        COMPACT_SERIALIZE_OFFSETS.owner.h2
+      )
+    );
+    const h2 = bigintFromBEBytes(
+      buf.slice(
+        COMPACT_SERIALIZE_OFFSETS.owner.h2,
+        COMPACT_SERIALIZE_OFFSETS.asset
+      )
+    );
+    const compressedOwner = { h1, h2 };
+    const owner = StealthAddressTrait.decompress(compressedOwner);
+
+    const asset = AssetTrait.deserializeCompact(
+      buf.slice(
+        COMPACT_SERIALIZE_OFFSETS.asset,
+        COMPACT_SERIALIZE_OFFSETS.nonce
+      )
+    );
+    const nonce = bigintFromBEBytes(
+      buf.slice(
+        COMPACT_SERIALIZE_OFFSETS.nonce,
+        COMPACT_SERIALIZE_OFFSETS.value
+      )
+    );
+    const value = bigintFromBEBytes(
+      buf.slice(COMPACT_SERIALIZE_OFFSETS.value, COMPACT_SERIALIZE_OFFSETS.end)
+    );
+
     return { owner, nonce, asset, value };
   }