+++ packages/sdk/src/primitives/types.ts
@@ -1,7 +1,8 @@
 import { CanonAddress, CompressedStealthAddress } from "../crypto";
 import { JoinSplitInputs, MerkleProofInput, SolidityProof } from "../proof";
 import { IncludedNote, Note } from "./note";
-import { Asset, EncodedAsset } from "./asset";
+import { EncodedAsset } from "./asset";
+import { HybridCiphertext } from "@nocturne-xyz/crypto-utils";
 
 export const BN254_SCALAR_FIELD_MODULUS =
   21888242871839275222246405745257275088548364400416034343698204186575808495617n;
@@ -38,16 +39,10 @@ export interface OperationResult {
   numRefunds: bigint;
 }
 
-export interface EncryptedNote {
-  owner: CompressedStealthAddress;
-  encappedKey: bigint;
-  encryptedNonce: bigint;
-  encryptedValue: bigint;
-}
+export type EncryptedNote = HybridCiphertext;
 
 export interface IncludedEncryptedNote extends EncryptedNote {
   merkleIndex: number;
-  asset: Asset;
   commitment: bigint;
 }